requirements.txt
feedparser
google-generativeai>=0.5.0
python-dotenv
fastapi
uvicorn[standard]
beautifulsoup4
requests
lxml